[easy] move file metadata write method.
This is required to break circular dependencies.
